INS-30: Move jet functions to a separate file.
